backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
sqlalchemy==2.0.36
pydantic[email]==2.10.4
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
langchain==0.3.14
langchain-anthropic==0.3.4
websockets==14.1
python-dotenv==1.0.1
pytest==8.3.4
hypothesis==6.122.3
httpx==0.28.1
